♻️ refactor: reorganize system architecture and module structure
- Move process execution from system/infra/ to system/ level
- Move UI components from system/infra/ui/ to system/ui/
- Move capability implementations (caddy, podman) to engine/capabilities/
- Flatten infra module hierarchy for better code organization
- Update all imports to reflect new module paths
- System module now directly exports fs, process, and ui submodules
