+++ workspace/src/core/context.rs
@@ -1,4 +1,4 @@
-use crate::engine::{Error, models::runtime::Runtime, system::infra::ui::Ui};
+use crate::engine::{Error, models::runtime::Runtime, system::ui::Ui};
 
 pub type AppResult<T> = Result<T, Error>;
