+++ workspace/src/engine/capabilities/podman/mod.rs
@@ -1,4 +1,23 @@
-/// Podman capability stub
-pub fn name() -> &'static str {
-    "podman"
+use crate::engine::{ErrorCode, Result, models::runtime::RunMode, system::process::Process};
+
+// ── Podman passthrough wrappers ─────────────────────────────────────────────
+pub async fn run(mode: RunMode, args: Vec<String>) -> Result<()> {
+    let proc = Process::new("podman").set_args(args).set_mode(mode);
+
+    if !proc.binary_exists() {
+        return Err(ErrorCode::PodmanMissing
+            .error()
+            .with_context("binary", "podman not found in PATH"));
+    }
+    proc.run_async().await?.emit()
+}
+
+pub async fn compose(mode: RunMode, args: Vec<String>) -> Result<()> {
+    let proc = Process::new("podman-compose").set_args(args).set_mode(mode);
+    if !proc.binary_exists() {
+        return Err(ErrorCode::PodmanComposeMissing
+            .error()
+            .with_context("binary", "podman-compose not found in PATH"));
+    }
+    proc.run_async().await?.emit()
 }
