--- workspace/src/engine/system/infra/mod.rs
@@ -1,4 +1,0 @@
-pub mod caddy;
-pub mod podman;
-pub mod process;
-pub mod ui;
